Move Flatbuffers marshalling to separate file
